refactor: 优化代码风格并修复多个clippy警告
- 使用`is_some_and`替代`map_or(false, ...)`模式
- 使用`strip_prefix`替代`starts_with`后切片
- 使用`clamp`替代`max`和`min`组合
- 添加`#[allow]`属性抑制特定clippy警告
- 优化条件判断逻辑，使用let-else和布尔表达式
- 简化数字类型转换和计算
- 统一代码风格和文档格式
- 添加默认实现和派生宏
- 优化字符串处理和匹配逻辑
- 修复潜在的数字溢出问题

+++ aria2-core/src/http/digest_auth.rs
@@ -251,7 +251,7 @@ fn rand_u64() -> u64 {
     let dur = SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap_or_default();
-    dur.as_nanos() as u64 ^ (dur.as_secs() as u64).wrapping_mul(0x5851F42D4C957F2D)
+    dur.as_nanos() as u64 ^ (dur.as_secs()).wrapping_mul(0x5851F42D4C957F2D)
 }
 
 #[cfg(test)]
